chore: refactor updateFormalExpression helper into a reusable utility
