+++ src/provider/bpmn/properties/AdHocCompletionProps.js
@@ -6,9 +6,8 @@ import {
   TextFieldEntry,
 } from '@bpmn-io/properties-panel';
 
-import { createElement } from '../../../utils/ElementUtil';
-
 import { useService } from '../../../hooks';
+import { createOrUpdateFormalExpression } from '../../../utils/FormalExpressionUtil';
 
 /**
  * @typedef { import('@bpmn-io/properties-panel').EntryDefinition } Entry
@@ -52,9 +51,13 @@ function CompletionCondition(props) {
   };
 
   const setValue = (value) => {
-    return commandStack.execute(
-      'element.updateModdleProperties',
-      updateFormalExpression(element, 'completionCondition', value, bpmnFactory)
+    return createOrUpdateFormalExpression(
+      element,
+      getBusinessObject(element),
+      'completionCondition',
+      value,
+      bpmnFactory,
+      commandStack
     );
   };
 
@@ -97,60 +100,4 @@ function CancelRemainingInstances(props) {
     getValue,
     setValue,
   });
-}
-
-// helper ////////////////////////////
-
-// formal expression /////////////////
-
-/**
- * updateFormalExpression - updates a specific formal expression
- *
- * @param {djs.model.Base} element
- * @param {string} propertyName
- * @param {string} newValue
- * @param {BpmnFactory} bpmnFactory
- */
-function updateFormalExpression(element, propertyName, newValue, bpmnFactory) {
-  const businessObject = getBusinessObject(element);
-  const expressionProps = {};
-
-  if (!newValue) {
-
-    // remove formal expression
-    expressionProps[propertyName] = undefined;
-
-    return {
-      element,
-      moddleElement: businessObject,
-      properties: expressionProps,
-    };
-  }
-
-  const existingExpression = businessObject.get(propertyName);
-  if (!existingExpression) {
-
-    // add formal expression
-    expressionProps[propertyName] = createElement(
-      'bpmn:FormalExpression',
-      { body: newValue },
-      businessObject,
-      bpmnFactory
-    );
-
-    return {
-      element,
-      moddleElement: businessObject,
-      properties: expressionProps,
-    };
-  }
-
-  // edit existing formal expression
-  return {
-    element,
-    moddleElement: existingExpression,
-    properties: {
-      body: newValue,
-    },
-  };
 }
