+++ src/provider/zeebe/properties/AdHocCompletionProps.js
@@ -4,11 +4,9 @@ import {
   isFeelEntryEdited,
 } from '@bpmn-io/properties-panel';
 
-import { createElement } from '../../../utils/ElementUtil';
-
 import { useService } from '../../../hooks';
-
 import { FeelEntryWithVariableContext } from '../../../entries/FeelEntryWithContext';
+import { createOrUpdateFormalExpression } from '../../../utils/FormalExpressionUtil';
 
 /**
  * @typedef { import('@bpmn-io/properties-panel').EntryDefinition } Entry
@@ -47,9 +45,13 @@ function CompletionCondition(props) {
   };
 
   const setValue = (value) => {
-    return commandStack.execute(
-      'element.updateModdleProperties',
-      updateFormalExpression(element, 'completionCondition', value, bpmnFactory)
+    return createOrUpdateFormalExpression(
+      element,
+      getBusinessObject(element),
+      'completionCondition',
+      value,
+      bpmnFactory,
+      commandStack
     );
   };
 
@@ -62,60 +64,4 @@ function CompletionCondition(props) {
     setValue,
     debounce,
   });
-}
-
-// helper ////////////////////////////
-
-// formal expression /////////////////
-
-/**
- * updateFormalExpression - updates a specific formal expression
- *
- * @param {djs.model.Base} element
- * @param {string} propertyName
- * @param {string} newValue
- * @param {BpmnFactory} bpmnFactory
- */
-function updateFormalExpression(element, propertyName, newValue, bpmnFactory) {
-  const businessObject = getBusinessObject(element);
-  const expressionProps = {};
-
-  if (!newValue) {
-
-    // remove formal expression
-    expressionProps[propertyName] = undefined;
-
-    return {
-      element,
-      moddleElement: businessObject,
-      properties: expressionProps,
-    };
-  }
-
-  const existingExpression = businessObject.get(propertyName);
-  if (!existingExpression) {
-
-    // add formal expression
-    expressionProps[propertyName] = createElement(
-      'bpmn:FormalExpression',
-      { body: newValue },
-      businessObject,
-      bpmnFactory
-    );
-
-    return {
-      element,
-      moddleElement: businessObject,
-      properties: expressionProps,
-    };
-  }
-
-  // edit existing formal expression
-  return {
-    element,
-    moddleElement: existingExpression,
-    properties: {
-      body: newValue,
-    },
-  };
 }
